requirements.txt
Adafruit-Blinka==5.9.0
adafruit-circuitpython-neopixel==6.0.0
adafruit-circuitpython-pypixelbuf==2.2.0
Adafruit-PlatformDetect==2.23.0
Adafruit-PureIO==1.1.8
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.62.0
fastrand==1.2.4
h11==0.11.0
idna==2.10
Pillow==8.1.0
pkg-resources==0.0.0
pydantic==1.7.3
pyftdi==0.52.0
pyserial==3.5
pyusb==1.1.0
requests==2.25.1
rpi-ws281x==4.2.5
RPi.GPIO==0.7.0
starlette==0.13.6
sysv-ipc==1.0.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.1
